he-plotly/requirements.txt
requests
click
helium-commander
dpath
requests
apscheduler
plotly
